+++ src/map/index.ts
@@ -5,6 +5,8 @@ import getWebpackConnections from "./next_js/getWebpackConnections.js";
 import getFetchInstances from "./next_js/getFetchInstances.js";
 import resolveFetch from "./next_js/resolveFetch.js";
 import interactive from "./next_js/interactive.js";
+import { existsSync, readFileSync } from "fs";
+import { Chunks } from "../utility/interfaces.js";
 
 const availableTech = {
     next: "Next.JS",
@@ -15,12 +17,12 @@ const availableFormats = {
 };
 
 const map = async (
-    directory,
-    output,
-    formats,
-    tech,
-    list,
-    interactive_mode
+    directory: string,
+    output: string,
+    formats: string[],
+    tech: string,
+    list: boolean,
+    interactive_mode: boolean
 ) => {
     console.log(chalk.cyan("[i] Running 'map' module"));
 
@@ -57,16 +59,33 @@ const map = async (
     }
 
     if (tech === "next") {
-        let chunks = await getWebpackConnections(directory, output, formats);
+        let chunks: Chunks;
 
-        // now, iterate through them, and check fetch instances
-        chunks = await getFetchInstances(chunks, output, formats);
+        let allOutputFilesAvailable = true;
+        Object.keys(availableFormats).map((key, value) => {
+            if (!existsSync(`${output}.${key}`)) {
+                allOutputFilesAvailable = false;
+            }
+        });
+
+        if (!allOutputFilesAvailable) {
+            // skip regeneration if output file already exists
+            chunks = await getWebpackConnections(directory, output, formats);
+
+            // now, iterate through them, and check fetch instances
+            chunks = await getFetchInstances(chunks, output, formats);
+        } else {
+            // read the JSON file, and load the value
+            chunks = JSON.parse(
+                readFileSync(`${output}.json`, { encoding: "utf8" })
+            );
+        }
 
         // resolve fetch once you've got all
         await resolveFetch(chunks, directory, formats);
 
         if (interactive_mode) {
-            await interactive(chunks);
+            await interactive(chunks, `${output}.json`);
         }
     }
 };