+++ src/map/next_js/interactive.ts
@@ -12,9 +12,10 @@ export interface State {
     commandHistory: string[];
     commandHistoryIndex: number;
     writeimports: boolean;
+    mapFile: string;
 }
 
-const interactive = async (chunks: Chunks) => {
+const interactive = async (chunks: Chunks, map_file: string) => {
     const state = {
         chunks,
         lastCommandStatus: true,
@@ -23,7 +24,8 @@ const interactive = async (chunks: Chunks) => {
         funcWriteFile: undefined,
         commandHistory: [],
         commandHistoryIndex: -1,
-        writeimports: true,
+        writeimports: false,
+        mapFile: map_file,
     };
 
     const ui = createUI();